Adds MeasureTable component to split measures
