+++ frontend/ui/subview.jsx
@@ -1,6 +1,5 @@
 import _ from 'lodash';
 import moment from 'moment';
-import numeral from 'numeral';
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { connect } from 'react-redux';
@@ -8,7 +7,8 @@ import { Button, ButtonGroup } from 'reactstrap';
 import Loading from './loading';
 import SubViewNav from './subviewnav';
 import { semVerCompare } from '../version';
-import { ORDER } from '../schema';
+import { CRASH_MEASURE_ORDER } from '../schema';
+import MeasureTable from './measureTable';
 
 const mapStateToProps = (state, ownProps) => {
   const { channel, platform } = ownProps.match.params;
@@ -41,26 +41,6 @@ const mapStateToProps = (state, ownProps) => {
   }
 
   return { measures: [] };
-};
-
-const getReleaseValue = (release, countType, timeWindow) => {
-  const property =
-    timeWindow === 'adjusted'
-      ? `adjusted${_.capitalize(countType)}`
-      : countType;
-
-  if (_.isUndefined(release) || _.isUndefined(release[property])) {
-    return 'N/A';
-  }
-
-  return (
-    <span
-      title={
-        countType === 'count' ? numeral(release[property]).format('0,0') : null
-      }>
-      {numeral(release[property]).format('0.00a')}
-    </span>
-  );
 };
 
 export class SubViewComponent extends React.Component {
@@ -97,18 +77,17 @@ export class SubViewComponent extends React.Component {
   }
 
   splitByOrder(measures) {
-    const main = [];
-
-    for (let i = 0; i < ORDER.length; i += 1) {
-      main.push(measures.find(element => element.name === ORDER[i]));
-    }
-
-    const others = _.orderBy(_.difference(measures, main), ['name']);
+    const crashMeasures = CRASH_MEASURE_ORDER.map(measureName =>
+      measures.find(element => element.name === measureName)
+    );
+    const otherMeasures = _.orderBy(_.difference(measures, crashMeasures), [
+      'name',
+    ]);
 
     this.setState({
       measures: {
-        main,
-        others,
+        crashMeasures,
+        otherMeasures,
       },
     });
   }
@@ -179,107 +158,24 @@ export class SubViewComponent extends React.Component {
                   </p>
                 </center>
               </div>
-              <h3>Crash Measures</h3>
-              <table className="table">
-                <thead>
-                  <tr>
-                    <th>Measure</th>
-                    {this.props.versions.map(version => (
-                      <th key={`th-${version}`}>{version}</th>
-                    ))}
-                    <th>Last updated</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {this.state.measures.main.map(measure => (
-                    <tr key={measure.name}>
-                      <td>
-                        <a
-                          href={`#/${this.state.channel}/${
-                            this.state.platform
-                          }/${measure.name}`}>
-                          {measure.name}
-                        </a>
-                      </td>
-                      {this.props.versions.map(versionStr => (
-                        <td
-                          key={`${measure.name}-${versionStr}`}
-                          title={
-                            versionStr.includes('.')
-                              ? `Events when ${versionStr} was latest`
-                              : `All values from all point releases during selected time window when ${versionStr} was latest`
-                          }>
-                          {getReleaseValue(
-                            measure.versions.find(
-                              release => release.version === versionStr
-                            ),
-                            this.state.countType,
-                            this.state.timeWindow
-                          )}
-                        </td>
-                      ))}
-                      <td al>
-                        {measure.lastUpdated
-                          ? moment(measure.lastUpdated).fromNow()
-                          : 'N/A'}
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-              <h3>Others</h3>
-              <table className="table">
-                <thead>
-                  <tr>
-                    <th>Measure</th>
-                    {this.props.versions.map(version => (
-                      <th key={`th-${version}`}>{version}</th>
-                    ))}
-                    <th>Last updated</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {this.state.measures.others.map(measure => (
-                    <tr key={measure.name}>
-                      <td>
-                        <a
-                          href={`#/${this.state.channel}/${
-                            this.state.platform
-                          }/${measure.name}`}>
-                          {measure.name}
-                        </a>
-                      </td>
-                      {this.props.versions.map(versionStr => (
-                        <td
-                          key={`${measure.name}-${versionStr}`}
-                          title={
-                            versionStr.includes('.')
-                              ? `Events when ${versionStr} was latest`
-                              : `All values from all point releases during selected time window when ${versionStr} was latest`
-                          }>
-                          {getReleaseValue(
-                            measure.versions.find(
-                              release => release.version === versionStr
-                            ),
-                            this.state.countType,
-                            this.state.timeWindow
-                          )}
-                        </td>
-                      ))}
-                      <td al>
-                        {measure.lastUpdated
-                          ? moment(measure.lastUpdated).fromNow()
-                          : 'N/A'}
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
+              <MeasureTable
+                title="Crash Measures"
+                measures={this.state.measures.crashMeasures}
+                versions={this.props.versions}
+                subviewState={this.state}
+              />
+              <MeasureTable
+                title="Other Measures"
+                measures={this.state.measures.otherMeasures}
+                versions={this.props.versions}
+                subviewState={this.state}
+              />
               {this.props.latestReleaseAge &&
                 this.props.latestReleaseAge < 86400 && (
                   <p className="text-danger">
                     <center>
-                      Latest release is new ({moment
+                      Latest release is new (
+                      {moment
                         .duration(this.props.latestReleaseAge, 'seconds')
                         .humanize()}{' '}
                       in field), numbers should be viewed with skepticism
